style: add schedule 컴포넌트 생성 및 코드 이동
